+++ courses/coding-in-C++/Lab_2/myLab2sec4/drink_utils.hpp
@@ -27,8 +27,8 @@ public:
 
 private:
 	std::string name;
-	int sugar;
-	int temperature;
+	uint8_t sugar;
+	uint8_t temperature;
 	bool withMilk;
 };
